+++ src/attr/converters.py
@@ -1,0 +1,23 @@
+"""
+Commonly useful converters.
+"""
+
+from __future__ import absolute_import, division, print_function
+
+
+def optional(converter):
+    """
+    A converter that allows an attribute to be optional. An optional attribute
+    is one which can be set to ``None``.
+
+    :param converter: the converter that is used for non-``None`` values.
+
+    ..  versionadded:: 17.1.0 *convert* can be optional
+    """
+
+    def optional_converter(val):
+        if val is None:
+            return None
+        return converter(val)
+
+    return optional_converter